requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
mysqlclient==2.2.7
sqlparse==0.5.5
firebase-admin==5.2.0
python-dotenv==1.0.0
django-cors-headers==3.13.0
gunicorn==20.1.0
cryptography==41.0.3
whitenoise==6.4.0
